// calculatormaatwerk.client/src/app/models/windowsill.model.ts
export interface WindowSill {
  id: number;
  name: string;
  referenceCoeck: string;
  referenceStore: string;
  store: string;
  color: string;
  picture: string;
  gsrn: number;
  isFourSidesOnlay: boolean;
  minLength: number;
  maxLength: number;
  minWidth: number;
  maxWidth: number;
  type: string;
  price: number;
  language: string;
  isExclBtw: boolean;
}